re-organized db adapter code. tables have their own respective js files.
